Rework scaling logic, OK timesheets are only minimally modified
The scaling logic seems to work a whole lot better if the scaling factor
is calculated on a "round to nearest (but never zero)" basis.

Also, timesheets which have been corrected should not be scaled again,
at best they should be rounded to the nearest numbers. They can then
pass through the FullWeekFilter (to be implemented) so that they are
ready for entry.