Add ElasticFilter
This filter will search for a time entry tagged "Elastic" for items
where it's not really clear what it belongs to. The filter will then
check what other (non-readonly) entries there are in that day and fairly
distribute the "elastic time" across them.

This might introduce timestamps with irregular fractions of a minute, so
it is recommended to pass the result through a rounding filter.