Subtract filter checks only entries in subtracted sheet
Before the filter would go over each day and entry in the result sheet
and try to find a similar entry in the subtracted sheet. If the two
didn't match then there was an error. This would give errors if the
result sheet contained (empty) days which didn't occur in the subtracted
sheet.

Reversing the operation makes it so that the subtracted sheet can
contain fewer items than the result sheet. If the subtracted sheet
contains an "Elastic" tag then it will be ignored, as it is assumed that
this tag will be absorbed into other tags of the result sheet.